requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.14.0
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.0
django-markdown2==0.3.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
gunicorn==20.1.0
markdown2==2.4.8
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
whitenoise==6.4.0
